requirements.txt
flask==2.0.1
flask-cors==3.0.10
python-dotenv==0.19.0
requests==2.26.0
tinydb==4.8.0
textrazor==1.4.0
elasticsearch==7.15.0
transformers==4.30.0
--find-links https://download.pytorch.org/whl/torch_stable.html
torch==2.0.0+cu118
torchvision==0.15.1+cu118
torchaudio==2.0.1+cu118
sentencepiece==0.1.99
accelerate==0.20.3
bitsandbytes-windows==0.37.5
werkzeug==2.2.2